# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/admin/group": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询或创建单条信息...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "查询或创建单条信息",
                "parameters": [
                    {
                        "description": "查询或创建单条信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GroupRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            }
        },
        "/admin/groups": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询多条信息...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "查询多条信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GroupListRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建单条信息...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "创建单条信息",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GroupRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "批量删除...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "批量删除",
                "parameters": [
                    {
                        "description": "批量删除",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BulkDelete"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPBulkSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "批量更新...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "批量更新",
                "parameters": [
                    {
                        "description": "批量更新",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BulkUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPBulkSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            }
        },
        "/admin/groups/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询单条信息...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "查询单条信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GroupRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除单条信息...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "删除单条信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPDeleteSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新单条信息...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "更新单条信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新单条信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GroupRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Auth": {
            "type": "object",
            "properties": {
                "authCode": {
                    "description": "鉴权识别码",
                    "type": "string"
                },
                "authName": {
                    "description": "鉴权名称",
                    "type": "string"
                },
                "authType": {
                    "description": "鉴权类型",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "expireTime": {
                    "description": "过期时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "isEnabled": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user": {
                    "description": "用户",
                    "type": "object",
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "string"
                },
                "verifyTime": {
                    "description": "认证时间",
                    "type": "string"
                }
            }
        },
...
        "model.BulkDelete": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.BulkUpdate": {
            "type": "object",
            "required": [
                "ids",
                "obj"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "obj": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "model.Group": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "leader": {
                    "description": "队长",
                    "type": "object",
                    "$ref": "#/definitions/model.User"
                },
                "leaderId": {
                    "description": "队长ID",
                    "type": "string"
                },
                "members": {
                    "description": "成员",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserGroup"
                    }
                },
                "name": {
                    "description": "组名称",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "model.GroupListRes": {
            "type": "object",
            "required": [
                "code",
                "data",
                "message",
                "pagin"
            ],
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Group"
                    }
                },
                "message": {
                    "description": "提示",
                    "type": "string"
                },
                "pagin": {
                    "type": "object",
                    "$ref": "#/definitions/model.PaginRes"
                }
            }
        },
        "model.GroupRes": {
            "type": "object",
            "required": [
                "code",
                "data",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/model.Group"
                },
                "message": {
                    "description": "提示",
                    "type": "string"
                }
            }
        },
        "model.HTTPBulkSuccess": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "提示",
                    "type": "string"
                }
            }
        },
        "model.HTTPDeleteSuccess": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据",
                    "type": "string"
                },
                "message": {
                    "description": "提示",
                    "type": "string"
                }
            }
        },
        "model.HTTPError": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "errors": {
                    "description": "详细错误信息",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "提示",
                    "type": "string"
                }
            }
        },
...
        "model.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "子菜单",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent": {
                    "description": "父菜单",
                    "type": "object",
                    "$ref": "#/definitions/model.Menu"
                },
                "parentId": {
                    "description": "父ID",
                    "type": "string"
                },
                "path": {
                    "description": "菜单路径",
                    "type": "string"
                },
                "roles": {
                    "description": "角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "model.PaginRes": {
            "type": "object",
            "required": [
                "page",
                "size",
                "total"
            ],
            "properties": {
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "menus": {
                    "description": "菜单",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "users": {
                    "description": "用户",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "auths": {
                    "description": "帐号",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Auth"
                    }
                },
                "avatar": {
                    "description": "昵称",
                    "type": "string"
                },
                "birthday": {
                    "description": "生日",
                    "type": "string"
                },
                "bloodType": {
                    "description": "血型(ABO)",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "friends": {
                    "description": "友",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                },
                "gender": {
                    "description": "性别",
                    "type": "integer"
                },
                "groups": {
                    "description": "用户组",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserGroup"
                    }
                },
                "height": {
                    "description": "身高(cm)",
                    "type": "number"
                },
                "homepage": {
                    "description": "个人主页",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "intro": {
                    "description": "简介",
                    "type": "string"
                },
                "lives": {
                    "description": "生活轨迹",
                    "type": "string"
                },
                "luckyNumbers": {
                    "description": "幸运数字",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机",
                    "type": "string"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "notice": {
                    "description": "备注",
                    "type": "string"
                },
                "roles": {
                    "description": "角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "score": {
                    "description": "积分",
                    "type": "integer"
                },
                "tags": {
                    "description": "标签",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userNo": {
                    "description": "编号",
                    "type": "integer"
                }
            }
        },
        "model.UserGroup": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "group": {
                    "description": "组",
                    "type": "object",
                    "$ref": "#/definitions/model.Group"
                },
                "groupId": {
                    "description": "组ID",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "joinTime": {
                    "description": "加入时间",
                    "type": "string"
                },
                "level": {
                    "description": "级别",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user": {
                    "description": "用户",
                    "type": "object",
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}